pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meta-agents"
version = "0.1.0"
description = "Meta-Agentic SDK - Declarative configuration for AI agents with Hermes backend"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
 {name = "Meta-Agentic Team"}
]
keywords = ["ai", "agents", "hermes", "sdk", "llm"]
classifiers = [
 "Development Status :: 3 - Alpha",
 "Intended Audience :: Developers",
 "License :: OSI Approved :: MIT License",
 "Programming Language :: Python :: 3",
 "Programming Language :: Python :: 3.11",
 "Programming Language :: Python :: 3.12",
]
dependencies = [
 "pydantic>=2.0.0",
 "typer>=0.9.0",
 "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
 "pytest>=7.0.0",
 "pytest-cov>=4.0.0",
 "black>=23.0.0",
 "ruff>=0.1.0",
 "mypy>=1.0.0",
]

[project.scripts]
meta-sdk = "meta_agents.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["meta_agents*"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
